+++ wcfsetup/install/files/lib/page/AttachmentPage.class.php
@@ -11,7 +11,7 @@ use wcf\util\FileReader;
 /**
  * Shows an attachment.
  * 
- * @author	Marcel Werk, Joshua Rüsweg
+ * @author	Joshua Ruesweg, Marcel Werk
  * @copyright	2001-2015 WoltLab GmbH
  * @license	GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
  * @package	com.woltlab.wcf